stuff - remove minifiers - move compressors to global scope instead of behind `CompressionTechnique` - fix Swift 6.0 error - move compression & decompression to their own files - remove `DecompressionError.unsupportedOperation` parameter - remove `@inlinable` annotations - support noncopyable types in protocols
