+++ Sources/LZ/brotli/Brotli+Decompress.swift
@@ -1,0 +1,14 @@
+
+import SwiftCompressionUtilities
+
+extension Brotli { // TODO: finish
+    /// - Parameters:
+    ///   - data: Collection of bytes to decompress.
+    ///   - closure: Logic to execute when a byte is decompressed.
+    /// - Complexity: O(_n_) where _n_ is the length of `data`.
+    public func decompress(
+        data: some Collection<UInt8>,
+        closure: (UInt8) -> Void
+    ) {
+    }
+}
