+++ Sources/LZ/deflate/DEFLATE+Decompress.swift
@@ -1,0 +1,10 @@
+
+import SwiftCompressionUtilities
+
+extension Deflate {
+    public static func decompress(
+        data: [UInt8]
+    ) -> [UInt8] {
+        return data
+    }
+}
