+++ Sources/SwiftCompressionUtilities/CompressionTechnique.swift
@@ -50,3 +50,38 @@ extension CompressionTechnique {
         return table
     }
 }
+
+#if compiler(>=6.2)
+// MARK: Inline Frequency tables
+extension CompressionTechnique {
+    /// Creates a universal frequency table from a sequence of raw bytes.
+    /// 
+    /// - Parameters:
+    ///   - data: Sequence of raw bytes.
+    /// - Returns: A universal frequency table.
+    /// - Complexity: O(_n_) where _n_ is the length of `data`.
+    public static func buildInlineFrequencyTable(data: some Sequence<UInt8>) -> [255 of Int] {
+        var table = [255 of Int](repeating: 0)
+        for byte in data {
+            table[Int(byte)] += 1
+        }
+        return table
+    }
+
+    /// Creates a universal frequency table from a character frequency dictionary.
+    /// 
+    /// - Parameters:
+    ///   - chars: A frequency table that represents how many times a character is present.
+    /// - Returns: A universal frequency table.
+    /// - Complexity: O(_n_) where _n_ is the sum of the `Character` utf8 lengths in `chars`.
+    public static func buildInlineFrequencyTable(chars: [Character:Int]) -> [255 of Int] {
+        var table = [255 of Int](repeating: 0)
+        for (char, freq) in chars {
+            for byte in char.utf8 {
+                table[Int(byte)] = freq
+            }
+        }
+        return table
+    }
+}
+#endif
