--- Sources/SwiftCompressionUtilities/ProgrammingLanguage.swift
@@ -1,7 +1,0 @@
-
-/// List of supported programming languages with compression techniques.
-public enum ProgrammingLanguage: Sendable {
-    case css
-    case javascript
-    case swift
-}
