+++ Sources/SwiftCompressionUtilities/errors/DecompressionError.swift
@@ -2,5 +2,5 @@
 /// SwiftCompression errors that can be thrown when decompressing data.
 public enum DecompressionError: Error {
     case malformedInput
-    case unsupportedOperation(String = "")
+    case unsupportedOperation
 }
