+++ Tests/DNATests/DNASingleBlockEncodingTestsTests.swift
@@ -14,7 +14,7 @@ import Testing
 struct DNASingleBlockEncodingTests {
     static let sequence:String = "TACTTGNCTAAAAGTACNATTGNCTAAGANTACACCGGCA"
     static let data:[UInt8] = [UInt8](sequence.utf8)
-    static let binary:[UInt8:[UInt8]] = CompressionTechnique.DNASingleBlockEncoding.compressBinary(data: data)
+    static let binary:[UInt8:[UInt8]] = DNASingleBlockEncoding.compressBinary(data: data)
 
     @Test func compressDNACSingleBlockEncodingPhase1() {
         #expect(Self.binary == [
@@ -30,7 +30,7 @@ struct DNASingleBlockEncodingTests {
     }
 
     @Test func compressDNACSingleBlockEncodingPhase2() {
-        let (result, controlBits):([UInt8], [UInt8]) = CompressionTechnique.DNASingleBlockEncoding.compressSBE(binaryData: Self.binary[65]!.prefix(7))
+        let (result, controlBits):([UInt8], [UInt8]) = DNASingleBlockEncoding.compressSBE(binaryData: Self.binary[65]!.prefix(7))
         #expect(result == [0, 1, 0])
         #expect(controlBits == [0])
     }
