--- Tests/SwiftCompressionTests/SwiftCompressionTests.swift
@@ -1,10 +1,0 @@
-
-#if compiler(>=6.0)
-
-import Testing
-@testable import SwiftCompression
-
-struct SwiftCompressionTests {
-}
-
-#endif
